Flip HttpClient.Metrics to extend HttpClientContext.Metrics instead of the other way around
